build nav section for admin panel

+++ routes/web.php
@@ -17,12 +17,12 @@ Route::get('/dashboard',[ShopController::class, 'index'])
 Route::resource('post',PostController::class);
 Route::middleware('admin.email')->group(function(){
 
-Route::get('/list_products',[ProductsController::class,'list_products'])->name('list_products');
-Route::get('/add_product',[ProductsController::class,'add_product'])->name('add_product');
-Route::post('/add_product',[ProductsController::class,'stor_product'])->name('products.store');
-Route::delete('/add_product/{id}',[ProductsController::class,'drop_products'])->name('delete_product');
-Route::get('/list_products/{product}/edit',[ProductsController::class, 'form_edit_product'])->name('form_edit_product');
-Route::patch('/list_products/{product}/edit',[ProductsController::class,'update_product'])->name('update_product');
+Route::get('/admin/list_products',[ProductsController::class,'list_products'])->name('list_products');
+Route::get('/admin/add_product',[ProductsController::class,'add_product'])->name('add_product');
+Route::post('/admin/add_product',[ProductsController::class,'stor_product'])->name('products.store');
+Route::delete('/admin/add_product/{id}',[ProductsController::class,'drop_products'])->name('delete_product');
+Route::get('/admin/list_products/{product}/edit',[ProductsController::class, 'form_edit_product'])->name('form_edit_product');
+Route::patch('/admin/list_products/{product}/edit',[ProductsController::class,'update_product'])->name('update_product');
 
 });
